+++ tests/func/regressions/issue52/test_issue52_table_caching.py
@@ -7,7 +7,7 @@ from parglare import Grammar, Parser
 def test_table_from_cache_different_than_calculated():
     this_folder = os.path.dirname(__file__)
     grammar_file = os.path.join(this_folder, 'grammar.pg')
-    table_file = os.path.join(this_folder, 'grammar.pgt')
+    table_file = os.path.join(this_folder, 'grammar.pgc')
     with contextlib.suppress(Exception):
         os.remove(table_file)